VER3-393 Massive refactor to allow docker deploy of back end
