+++ app/app/calculation_service/model/enums.py
