+++ app/uwsgi.ini
@@ -1,0 +1,3 @@
+[uwsgi]
+module = app.main
+callable = app
