+++ requirements.txt
@@ -2,3 +2,4 @@ numpy
 setuptools==28.8.0
 Flask==0.12.2
 flask_cors==3.0.3
+Flask-PyMongo==2.1.0
