+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='demoappback',
+    name='app',
     packages=find_packages(),
     include_package_data=True,
     install_requires=['flask', 'flask_cors', ],
